+++ philo/philo.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   philo.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: fbazaz <fbazaz@student.42.fr>              +#+  +:+       +#+        */
+/*   By: tiima <tiima@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/02 09:13:59 by fbazaz            #+#    #+#             */
-/*   Updated: 2024/06/03 12:01:08 by fbazaz           ###   ########.fr       */
+/*   Updated: 2024/06/05 16:31:06 by tiima            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -34,8 +34,8 @@ typedef struct s_philo
 	int				id;
 	int				eat_count;
 	int				is_eating;
-	u_int64_t		start;
-	u_int64_t		time_die;
+	unsigned long long		start;
+	unsigned long long		time_die;
 	pthread_mutex_t	*right_fork;
 	pthread_mutex_t	*left_fork;
 	pthread_mutex_t	lock;
@@ -47,9 +47,9 @@ typedef struct s_data
 	t_philo			*philo;
 	int				n_philo;
 	int				finish;
-	u_int64_t		t_die;
-	u_int64_t		t_eat;
-	u_int64_t		t_sleep;
+	unsigned long long		t_die;
+	unsigned long long		t_eat;
+	unsigned long long		t_sleep;
 	int				n_meals;
 	int				death_flag;
 	pthread_mutex_t	*forks;
@@ -73,7 +73,7 @@ void				*observer(void *arg);
 void				*routine(void *arg);
 int					init_threads(t_data *data);
 int					ft_atoi(char *str);
-u_int64_t			get_current_time(void);
+unsigned long long			get_current_time(void);
 int    				ft_usleep(size_t milliseconds);
