+++ philo/utils.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   utils.c                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: fbazaz <fbazaz@student.42.fr>              +#+  +:+       +#+        */
+/*   By: tiima <tiima@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/02 10:26:01 by fbazaz            #+#    #+#             */
-/*   Updated: 2024/06/03 12:06:14 by fbazaz           ###   ########.fr       */
+/*   Updated: 2024/06/05 16:29:09 by tiima            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ int    ft_usleep(size_t milliseconds)
     return (0);
 }
 
-u_int64_t	get_current_time(void)
+unsigned long long	get_current_time(void)
 {
 	struct timeval	tv;
